createConnect, with configurable shallowEqual
