+++ src/index.js
@@ -1,2 +1,3 @@
 export { default as Provider } from './components/Provider'
 export { default as connect } from './components/connect'
+export { default as createConnect } from './components/createConnect'
